requirements.txt
# --------- Deep Learning Frameworks --------- #
torch>=2.0.0           # PyTorch deep learning framework
torchvision>=0.15.0    # Utilities for PyTorch such as pretrained models and dataset loaders
lightning>=2.0.0       # PyTorch Lightning, a lightweight PyTorch wrapper
torchmetrics>=0.11.4   # Metrics for PyTorch

# --------- Configuration and Hyperparameter Optimization --------- #
hydra-core==1.3.2            # For managing configurations
hydra-colorlog==1.2.0        # Colorized logs with Hydra
hydra-optuna-sweeper==1.2.0  # Optuna plugin for Hydra for hyperparameter optimization

# --------- Logging and Monitoring --------- #
tensorboard             # TensorBoard for visualizing training metrics and profiles
# wandb                 # Weights and Biases for advanced experiment tracking and visualization
# neptune-client        # Neptune AI for managing and tracking ML experiments
# mlflow                # MLflow for managing the ML lifecycle, including experimentation, reproducibility, and deployment
# comet-ml              # Comet for tracking ML experiments, comparing models, and monitoring training in real time
# aim>=3.16.2           # Aim for experiment tracking, specifically version 3.16.2 or higher due to issue resolution (https://github.com/aimhubio/aim/issues/2550)

# --------- Data Manipulation and Analysis --------- #
pandas                 # Data manipulation and analysis
polars==0.19.7         # Fast DataFrame library
matplotlib             # Plotting and visualization library
japanize-matplotlib    # Japanese fonts for Matplotlib
nbformat>=4.2.0        # Notebook format for Jupyter

# --------- Utilities --------- #
rootutils              # For standardizing the project root setup
pre-commit             # Hooks for applying linters on commit
rich                   # For beautiful text formatting in the terminal
pytest                 # For running tests

# --------- Version Control for Data --------- #
dvc==3.23.0            # Data Version Control for tracking data and model versions (local)
dvc-s3==2.23.0         # DVC plugin for AWS S3 storage

# --------- System --------- #
sh                     # For running bash commands in some tests (linux/macos only)

# --------- Additional Libraries for Time Series and Stats --------- #
neuralprophet          # Time series forecasting
optuna                 # Hyperparameter optimization
statsmodels            # Statistical models

# --------- Jupyter Notebook --------- #
ipykernel              # Kernel for Jupyter notebooks

# --------- Additional Data Visualization Libraries --------- #
shap                   # SHAP (SHapley Additive exPlanations) for explaining machine learning model output
seaborn                # High-level interface for drawing attractive and informative statistical graphics
numpy                  # Numerical computing library

# --------- AWS SDK --------- #
boto3                  # AWS SDK for Python, for interacting with AWS services like S3, EC2, and more

# --------- Additional Utilities --------- #
kaggle                       # For interacting with Kaggle datasets and competitions
